workflow-agent/requirements.txt
kafka-python==2.0.2
marshmallow==3.20.1
jsonschema==4.19.0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
python-dateutil==2.8.2

